% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api1.R
\name{run_api1}
\alias{run_api1}
\title{Run API 1}
\usage{
run_api1(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run API 1
}
\examples{
\dontrun{
run_api1()
run_api1(swagger = TRUE, port = 8000)
}
}
