% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2.R
\name{run_api2}
\alias{run_api2}
\title{Run API 2}
\usage{
run_api2(...)
}
\arguments{
\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run API 2
}
\examples{
\dontrun{
run_api2()
run_api2(swagger = TRUE, port = 8000)
}
}
